+++ src/main/java/ru/yandex/practicum/filmorate/service/mpaRating/MpaRatingService.java
@@ -5,7 +5,7 @@ import ru.yandex.practicum.filmorate.model.MpaRating;
 import java.util.List;
 
 public interface MpaRatingService {
-    MpaRating getRatingFilm(Integer id);
+    MpaRating getRatingFilm(Long id);
 
     List<MpaRating> getRatingsFilms();
 }
